bot/pyproject.toml
[project]
name = "whatsapp-pdf-bot"
version = "0.1.0"
description = "Whatsapp PDF-Q&A bot with ChromaDB and Gemini"
authors = [
    {name = "AaravAsthana",email = "aaravasthana2308@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "google-genai (>=1.21.0,<2.0.0)",
    "flask (>=3.1.1,<4.0.0)",
    "pywa[flask] (>=2.11.0,<3.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "unstructured (>=0.17.2,<0.18.0)",
    "chromadb (>=1.0.13,<2.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-community (>=0.3.25,<0.4.0)",
    "llama-parse (>=0.6.34,<0.7.0)",
    "llama-cloud-services (>=0.6.34,<0.7.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
